Assignment 2
Insurance cost prediction using linear regression

In this assignment we're going to use information like a person's age, sex, BMI, no. of children and smoking habit to predict the price of yearly medical bills. This kind of model is useful for insurance companies to determine the yearly insurance premium for a person. The dataset for this problem is taken from Kaggle.

We will create a model with the following steps:

    Download and explore the dataset
    Prepare the dataset for training
    Create a linear regression model
    Train the model to fit the data
    Make predictions using the trained model